Do not try to delete tests migrations in older djangos

+++ run_tests.py
@@ -26,4 +26,5 @@ try:
         call_command('makemigrations', 'tests', verbosity=2 if '-v' in sys.argv else 0)
     call_command('test', names, failfast='-x' in sys.argv, verbosity=2 if '-v' in sys.argv else 1)
 finally:
-    shutil.rmtree('tests/migrations')
+    if django.VERSION >= (1, 7):
+        shutil.rmtree('tests/migrations')
